0.0.5
This update I created the transition from originally identifying cards by using strictly integers between 1-52 to using a Card Object, which contains four data members: ID,name,value, and SUIT.  The idea behind is at the beginning of initialization, an array for Card cards[52] will be launched automatically create the cards needed with the data members filled.  I was thinking to use this array as a way of identifying cards passed around between the players and the pile.

Also the pile object is now in existence and no longer built into the DKATGame class.  

This update goes as far as to now displaying the names of cards, which consist of the number and the letter of the suit ie) 4D - four of diamonds. 

Players are now able to accuse people of bluffing.  The system will run a check and determine if a bluff has been made.

I stopped for the moment being at what happens when a bluff is found or not found, i plan on figuring that out next.