0.1.0
This update consists of mode where you can play as only one of the characters, and the other players are controlled by logic, aka bots.  By default it is set, the user is in control of players[2].  The bots will decide whether or not to call out bluffs, and will bluff on their own, if they don't have good enough hands of their own.

All information that is displayed in the command prompt, that shouldn't be displayed normally is marked with five "-" before each line.

The organization of cards to be displayed will now be displayed in a value order as opposed to card id order.

There is now an ending sequence, although not that extravagant. 

At this point what's next to be done is to connect the code with a GUI, and then get a server running with the logic part of the code, in order to allow multi player connections.