The GUI LAYOUT
I have a window working for me, but I'm still getting used to the win32 native gui, that's based in C.  I'm definitely learning a lot from this process, I knew this would be one of the most difficult parts of this project.  It's just a matter of time, I know I can complete this task.  It will all be worth it once I get this guy running.