Add files via upload
This installment consists of a fully functional client.

This GUI only supports multiplayer functionality among other clients; no single card game playing.

There is a message box that displays game messages from a queue.  There is a connect button at the top, which requires the server's IP and port to connect to.  The cards are displayed for you on the right side of the window, large enough to fit all 52 cards.  You will also see the amount of cards each player has, as well as how many cards have accumulated in the pile.